added another irrklang repo + updated the ogre3d 1.7.4 freshly compiled in vs 2026, thanks to DylanCheetah for pointing me that PTR1 does not exist no more in the STD namespace

+++ Client/deps/OgreSDK_1.7-64/include/OGRE/asm_math.h
@@ -179,8 +179,8 @@ float asm_rsq( float r )
     } // returns r0
 
 #else
-
-	return 1. / sqrt( r );
+    // C4244
+	return (float)(1. / sqrt(r));
 
 #endif
 }
@@ -209,8 +209,8 @@ float apx_rsq( float r ) {
     } // returns r0
 
 #else
-
-	return 1. / sqrt( r );
+    // C4244
+	return (float)(1. / sqrt(r));
 
 #endif
 }
